refactor: Classify java files to Layered Directory Architecture
- Apply google code style
