+++ src/main/java/gift/controller/ProductController.java
@@ -1,12 +1,18 @@
-package gift;
+package gift.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import gift.model.Product;
+import gift.service.ProductOperation;
+import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Map;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/products")
@@ -14,7 +20,6 @@ public class ProductController {
 
     private final ProductOperation productOperation;
 
-    @Autowired
     public ProductController(ProductOperation productOperation) {
         this.productOperation = productOperation;
     }
@@ -41,11 +46,13 @@ public class ProductController {
     }
 
     @PutMapping("edit/{id}")
-    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody Product updatedProduct) {
+    public ResponseEntity<Product> updateProduct(@PathVariable Long id,
+        @RequestBody Product updatedProduct) {
         Product product = productOperation.getProductById(id);
         if (product != null) {
             updatedProduct.setId(id);
-            Product updated = productOperation.updateProduct(updatedProduct.getId(), updatedProduct);
+            Product updated = productOperation.updateProduct(updatedProduct.getId(),
+                updatedProduct);
             return ResponseEntity.ok(updated);
         }
         return ResponseEntity.notFound().build();
