+++ src/main/java/gift/model/Product.java
@@ -1,6 +1,7 @@
-package gift;
+package gift.model;
 
 public class Product {
+
     private Long id;
     private String name;
     private long price;
