+++ src/main/resources/templates/product-list.html
@@ -50,6 +50,7 @@
   }
 
 
+
 </script>
 </body>
 </html>
